removed extra property in this and made a separate module for button
